int64 should correspond to SQL BigInt

+++ SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/sqlparser/CalciteCompiler.java
@@ -185,13 +185,13 @@ public class CalciteCompiler implements IModule {
         CalciteSchema rootSchema = CalciteSchema.createRootSchema(false, false);
         rootSchema.add(catalog.schemaName, this.catalog);
         // Register new types
-        rootSchema.add("INT64", factory -> factory.createSqlType(SqlTypeName.INTEGER));
+        rootSchema.add("INT64", factory -> factory.createSqlType(SqlTypeName.BIGINT));
         rootSchema.add("FLOAT64", factory -> factory.createSqlType(SqlTypeName.DOUBLE));
         rootSchema.add("FLOAT32", factory -> factory.createSqlType(SqlTypeName.FLOAT));
         rootSchema.add("STRING", factory -> factory.createSqlType(SqlTypeName.VARCHAR));
         rootSchema.add("BOOL", factory -> factory.createSqlType(SqlTypeName.BOOLEAN));
         // TODO: not entirely correct
-        rootSchema.add("UINT64", factory -> factory.createSqlType(SqlTypeName.INTEGER));
+        rootSchema.add("UINT64", factory -> factory.createSqlType(SqlTypeName.BIGINT));
         Prepare.CatalogReader catalogReader = new CalciteCatalogReader(
                 rootSchema, Collections.singletonList(catalog.schemaName), this.typeFactory, connectionConfig);
